@@ -1,14 +1,17 @@
 module github.com/eoe2005/g
 
-go 1.17
+go 1.19
 
 require (
 	github.com/apolloconfig/agollo/v4 v4.2.1
 	github.com/bradfitz/gomemcache v0.0.0-20220106215444-fb4bf637b56d
 	github.com/gin-gonic/gin v1.8.1
 	github.com/go-redis/redis/v8 v8.11.5
+	github.com/google/uuid v1.3.0
 	github.com/segmentio/kafka-go v0.4.36
 	go.etcd.io/etcd/client/v3 v3.5.5
+	gopkg.in/gomail.v2 v2.0.0-20160411212932-81ebce5c23df
+	gopkg.in/yaml.v3 v3.0.1
 	gorm.io/driver/clickhouse v0.5.0
 	gorm.io/driver/mysql v1.4.3
 	gorm.io/driver/postgres v1.4.5
